refactor: move gconstant and genericdao out of biz package
