+++ biz/gmiddleware/ginmiddleware/token_blacklist.go
@@ -2,7 +2,7 @@ package ginmiddleware
 
 import (
 	"github.com/gin-gonic/gin"
-	"github.com/morehao/golib/biz/gconstant"
+	"github.com/morehao/golib/gconstant"
 	"github.com/morehao/golib/biz/gcontext/gincontext"
 	"github.com/morehao/golib/gcrypto"
 	"github.com/morehao/golib/gerror"
