+++ gconstant/api.go
