+++ gconstant/error_code.go
