+++ genericdao/base.go
