+++ genericdao/cond.go
